// src/enums/CursorTypes.ts
export enum CursorType {
    NORMAL = "normal.ani",
    NPC_TURN = "npc_turn.ani",
    ATTACK = "attack.ani",
    CANT_CAST = "cantcast.ani",
    CANT_OPEN = "cantopen.ani",
    CANT_TAKE = "canttake.ani",
    CAST = "cast.ani",
    DISTANCE_ATTACK = "distance_attack.ani",
    DOWN = "down.ani",
    LDOWN = "ldown.ani",
    LEFT = "left.ani",
    LUP = "lup.ani",
    NOWAY = "noway.ani",
    OPEN = "open.ani",
    RIGHT = "right.ani",
    RDOWN = "rdown.ani",
    RUP = "rup.ani",
    SAVE = "save.ani",
    TAKE = "take.ani",
    TALK = "talk.ani",
    UP = "up.ani",
    ANOTHER_LOCATION = "another_location.ani",
}